requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
bitarray==3.4.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.0
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
cytoolz==1.0.1
dnspython==2.7.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
Flask==3.1.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
flask-restx==1.3.0
frozenlist==1.6.0
h11==0.16.0
hexbytes==1.3.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonalias==0.1.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mongoengine==0.29.1
multidict==6.4.3
packaging==25.0
parsimonious==0.10.0
pluggy==1.5.0
propcache==0.3.1
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.12.1
pytest==8.3.5
python-bitcoinlib==0.12.2
python-dotenv==1.1.0
pytz==2025.2
pyunormalize==16.0.0
pywin32==310
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
rpds-py==0.24.0
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
toolz==1.0.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
web3==7.11.1
websockets==15.0
Werkzeug==3.1.3
yarl==1.20.0
